+++ lablib/processors/__init__.py
@@ -1,0 +1,13 @@
+from .processors import (
+    EffectsFileProcessor,
+    ColorProcessor,
+    RepoProcessor,
+    SlateProcessor,
+)
+
+__all__ = [
+    "EffectsFileProcessor",
+    "ColorProcessor",
+    "RepoProcessor",
+    "SlateProcessor",
+]
